Add anime search functionality to API
Introduced an anime search feature in the API. The 'Anime.ts' file was added to define the Anime type. This includes an async function to request a list of animes matching the given search term. Additionally, implemented the render function for AnimeTitle in the renderAnime function within the 'anime.tsx' file.