Refactor package.json commands and add README instructions
This commit simplifies the 'serve:extension:firefox' command in package.json and adds 'build' and 'bundle' commands for easier project management. It also introduces browser-specific settings and updates icon formats in the manifest.json file. Moreover, a README.md file is added, which includes detailed instructions on how to install, run and build the project.