Add AnimeRating component and render it
A new `AnimeRating` component has been created in the `components` directory, which displays an anime's score. Changes have also been made to the `main.css` and `anime.tsx` files to accommodate the inclusion of this new component in anime render. This enhancement allows users to see the ratings of animes within the extension.