Update package version and modify extension manifest settings
Bumped the version from 0.0.2 to 0.0.4 in package.json. The 'action' field in manifestTemplate.json was renamed to 'browser_actions' and manifest_version had its value dynamically determined. Also, redundant 'json-templates' dependency was removed.

With these changes, 'json-templates' is no longer required which simplifies our codebase. At the same time, the manifest file was updated to meet the requirements for extensions on browser platforms.